+++ src/App.scss
@@ -1,5 +1,9 @@
 @import-normalize;
 
+*, *::after, *::before {
+  box-sizing: border-box;
+}
+
 .App {
   color: white;
   background-color: #242738;
@@ -9,135 +13,4 @@
   flex-direction: column;
   align-items: center;
   justify-content: center;
-}
-
-
-form {
-  min-width: 180px;
-  max-width: 360px;
-  line-height: 1.5;
-}
-
-label, legend {
-  display: inline-block;
-  padding: 0;
-  margin-bottom: 4px;
-
-  font-size: 1rem;
-}
-
-input {
-  color: inherit;
-  line-height: inherit;
-  background: inherit;
-  transition: all 0.3s ease;
-
-  box-sizing: border-box;
-  width: 100%;
-  height: 2.8rem;
-  padding: 0;
-  padding-inline: 1rem;
-
-  background-color: hsl(215, 50%, 94%);
-  border: 1px solid rgb(175, 180, 200);
-  border-radius: 4px;
-
-  color: rgba(0, 0, 0, 0.8);
-  font-size: 1.5rem;
-
-  &:focus {
-    outline: none;
-    border-color: rgb(0, 128, 255);
-    box-shadow: 0 0 0 4px dodgerblue;
-  }
-}
-
-
-/* === Expiry Date === */
-fieldset {
-  /* reset */
-  border: none;
-  margin: 0;
-  padding: 0;
-  display: flex;
-  align-items: center;
-
-  > span {
-    margin-inline: 8px;
-  }
-
-  /* To continue using border-box, calc is used to include padding for width */
-  > input {
-    min-width: calc(2ch + 32px);
-    text-align: center;
-  }
-}
-
-
-/* === Security Code === */
-form > div:last-of-type {
-  @media (min-width: 360px) {
-    /* #cc_code */
-    > input {
-      min-width: calc(4ch + 32px);
-      text-align: center;
-    }
-  }
-}
-
-
-form {
-  display: grid;
-  align-items: end;
-  grid-gap: 16px;
-  column-gap: 32px; // Specifically for expiry + code row
-  grid-template-columns: 1fr;
-
-  box-sizing: border-box;
-  padding: 16px;
-
-  @media (min-width: 360px) {
-    grid-template-columns: repeat(10, 1fr);
-
-  > div:first-child, > button:last-child {
-    grid-column: span 10;
-  }
-  > fieldset {
-      grid-column: span 6;
-  }
-  > div:last-of-type {
-      grid-column: span 4;
-  }
-  }
-}
-
-button {
-  display: block;
-  cursor: pointer;
-  transition: all 0.3s ease;
-
-  margin-top: auto;
-  padding: 8px 32px;
-
-  width: 100%;
-  height: 2.8rem;
-
-  background-color: rgb(47, 104, 179);
-  border: none;
-  border-radius: 4px;
-
-  color: white;
-  font-size: 1rem;
-  font-weight: bold;
-  text-align: center;
-
-  &:active, &:hover {
-    background-color: dodgerblue;
-  }
-
-  &:focus {
-    outline: none;
-    background-color: dodgerblue;
-    box-shadow: 0 0 0 4px dodgerblue;
-  }
 }
